refactor(ffmpeg): remove multi-command mode and simplify command generation
