+++ charts/pre-api/values.stg.template.yaml
@@ -14,7 +14,6 @@ java:
     ENABLE_ENHANCED_PROCESSING: true
     ENABLE_NULL_DURATION_UPSERT: false
     ENABLE_MIGRATION_ADMIN_ENDPOINTS: true
-    ENABLE_MULTI_FFMPEG_COMMAND: false
     ENABLE_MIGRATED_DATA: true
   image: ${IMAGE_NAME}
   ingressHost: ${SERVICE_FQDN}
