+++ charts/pre-api/values.yaml
@@ -74,7 +74,6 @@ java: &javavalues #This is an anchor to reuse settings in Job mode.
     ENABLE_ENHANCED_PROCESSING: false
     ENABLE_NULL_DURATION_UPSERT: false
     ENABLE_MIGRATION_ADMIN_ENDPOINTS: false
-    ENABLE_MULTI_FFMPEG_COMMAND: false
     TASK_START_LIVE_EVENTS_BATCH_SIZE: 20
     TASK_START_LIVE_EVENTS_POLL_INTERVAL: 2000
     TASK_CLEANUP_LIVE_EVENTS_BATCH_SIZE: 3
